Add: plot only sequential log. Fix: Remove not functional option. Specification change: feed log directory instead of log file.

+++ plotlogseq.sh
@@ -77,7 +77,6 @@ fi
 
 cd "${LogDirectory}"
 
-
 for f in *.txt
 do
 	header=${TempPath}/${uuid}-`basename ${f%.*}-hd.txt`
@@ -96,37 +95,37 @@ do
 	Repeats=`grep 'Repeats(-n):' ${header} | cut -d ':' -f 2`
 	LBASectors=`sed -n '/LBAsects/ s/.*LBAsects=\([0-9][0-9]*\)/\1/p' ${header}`
 
-	if [[ -z  ${FileSize} ]]
+	if [[ -z ${FileSize} ]]
 	then
-		echo "${f}: Not access log, skip."
+		echo "${f}: Not access log, skip. FileSize not found."
 		rm ${header}
 		continue
 	fi
 
-	contain_plot_data=0
-	if [[ "${FillFile}x" == "yx" ]]
+	if [[ -z ${FillFile} ]]
 	then
-		contain_plot_data=1
+		echo "${f}: Not access log, skip. FillFile not found."
+		rm ${header}
+		continue
 	fi
 
-	if [[ -n ${Repeats} ]]
+	if [[ -z ${Repeats} ]]
 	then
-		if (( ${Repeats} > 0 ))
-		then
-			contain_plot_data=1
-		fi
-	else
-		Repeats=0
+		echo "${f}: Not access log, skip. Repeats not found."
+		rm ${header}
+		continue
 	fi
 
-	if [[ "${DoReadFile}x" != "nx" ]]
+	if [[ -z ${DoReadFile} ]]
 	then
-		contain_plot_data=1
+		echo "${f}: Not access log, skip. DoReadFile not found."
+		rm ${header}
+		continue
 	fi
 
-	if (( ${contain_plot_data} == 0 ))
+	if [[ ( "${FillFile}" != "y" ) && ( "${DoReadFile}" != "y" ) ]]
 	then
-		echo "${f}: Not access log, skip."
+		echo "${f}: Not contain sequential access log."
 		rm ${header}
 		continue
 	fi
@@ -164,18 +163,12 @@ do
 		DoDirectSequential='without O_DIRECT'
 	fi
 
-	DoDirectRandom='with O_DIRECT'
-	if ( echo ${DoDirect} | grep -q 'N' )
-	then
-		DoDirectRandom='without O_DIRECT'
-	fi
-
 	rm ${header}
 
 	GnuplotVarFile=${TempPath}/${uuid}-`basename ${f%.*}-gp.txt`
 	part_data_file=${TempPath}/${uuid}-`basename ${f%.*}-pd.txt`
 
-	if [[ "${FillFile}x" == "yx" ]]
+	if [[ "${FillFile}" == "y" ]]
 	then
 		cat << EOF > ${GnuplotVarFile}
 set title "${Model} ${CapacityGB}G bytes, sequential write\\n\
@@ -203,108 +196,7 @@ EOF
 		rm ${part_data_file}
 	fi
 
-	if (( ${Repeats} > 0 ))
-	then
-		ra_file=${TempPath}/${uuid}-`basename ${f%.*}-ra.txt`
-		sed -n '/^i,[[:space:]]/,/close/ {p}' ${f} | grep '^[[:space:]]*[0-9]' | sed -n 's/,//gp' > ${ra_file}
-		ra_file_size=`stat --format=%s ${ra_file}`
-		if (( ${ra_file_size} > 0 ))
-		then
-			ra_r_tspeed_at_png=${f%.*}-rr-ts_at.png
-			echo "${f}: ${ra_r_tspeed_at_png}: Plot random read transfer speed - access time."
-			cat << EOF > ${GnuplotVarFile}
-set title "${Model} ${CapacityGB}G bytes,\\n\
-plot reads of random read/write\\n\
-${RandomRWMinBytesKi}Ki to ${RandomRWMaxBytesKi}Ki bytes per one read() call, \
-${DoDirectRandom}\\ntransfer speed - access time"
-pointcolor="#00c000"
-set yrange [ ${RANDOM_TRANSFER_SPEED_MIN} : ${RANDOM_TRANSFER_SPEED_MAX} ] noreverse nowriteback
-EOF
-			grep 'r' "${ra_file}"  > ${part_data_file}
-			gnuplot -e "log_file=\"${part_data_file}\"; load \"${GnuplotVarFile}\"; \
-				load \"${my_dir}/random_tspeed_at.gnuplot\"; quit" \
-				> ${ra_r_tspeed_at_png}
-
-
-			ra_r_tlength_at_png=${f%.*}-rr-tl_at.png
-			echo "${f}: ${ra_r_tlength_at_png}: Plot random read transfer length - access time."
-			cat << EOF > ${GnuplotVarFile}
-set title "${Model} ${CapacityGB}G bytes,\\n\
-plot reads of random read/write\\n\
-${RandomRWMinBytesKi}Ki to ${RandomRWMaxBytesKi}Ki bytes per one read() call, \
-${DoDirectRandom}\\ntransfer length - access time"
-pointcolor="#00c000"
-EOF
-			gnuplot -e "log_file=\"${part_data_file}\"; load \"${GnuplotVarFile}\"; \
-				    load \"${my_dir}/random_tlength_at.gnuplot\"; quit" \
-				> ${ra_r_tlength_at_png}
-
-			ra_r_tspeed_tlength_png=${f%.*}-rr-ts_tl.png
-			echo "${f}: ${ra_r_tspeed_tlength_png}: Plot random read transfer speed - transfer length."
-			cat << EOF > ${GnuplotVarFile}
-set title "${Model} ${CapacityGB}G bytes,\\n\
-plot reads of random read/write\\n\
-${RandomRWMinBytesKi}Ki to ${RandomRWMaxBytesKi}Ki bytes per one read() call, \
-${DoDirectRandom}\\ntransfer speed - transfer length"
-pointcolor="#00c000"
-set yrange [ ${RANDOM_TRANSFER_SPEED_MIN} : ${RANDOM_TRANSFER_SPEED_MAX} ] noreverse nowriteback
-EOF
-			gnuplot -e "log_file=\"${part_data_file}\"; load \"${GnuplotVarFile}\"; \
-				    load \"${my_dir}/random_tspeed_tlength.gnuplot\"; quit" \
-				> ${ra_r_tspeed_tlength_png}
-
-			rm ${part_data_file}
-
-			ra_w_tspeed_at_png=${f%.*}-rw-ts_at.png
-			echo "${f}: ${ra_w_tspeed_at_png}: Plot random write transfer speed - access time."
-			cat << EOF > ${GnuplotVarFile}
-set title "${Model} ${CapacityGB}G bytes,\\n\
-plot writes of random read/write\\n\
-${RandomRWMinBytesKi}Ki to ${RandomRWMaxBytesKi}Ki bytes per one write() call, \
-${DoDirectRandom}\\ntransfer speed - access time"
-pointcolor="#ff0000"
-set yrange [ ${RANDOM_TRANSFER_SPEED_MIN} : ${RANDOM_TRANSFER_SPEED_MAX} ] noreverse nowriteback
-EOF
-			grep 'w' "${ra_file}" > ${part_data_file}
-			gnuplot -e "log_file=\"${part_data_file}\"; load \"${GnuplotVarFile}\"; \
-				    load \"${my_dir}/random_tspeed_at.gnuplot\"; quit" \
-				> ${ra_w_tspeed_at_png}
-
-			ra_w_tlength_at_png=${f%.*}-rw-tl_at.png
-			echo "${f}: ${ra_w_tlength_at_png}: Plot random write transfer length - access time."
-			cat << EOF > ${GnuplotVarFile}
-set title "${Model} ${CapacityGB}G bytes,\\n\
-plot writes of random read/write\\n\
-${RandomRWMinBytesKi}Ki to ${RandomRWMaxBytesKi}Ki bytes per one write() call, \
-${DoDirectRandom}\\ntransfer length - access time"
-pointcolor="#ff0000"
-EOF
-			gnuplot -e "log_file=\"${part_data_file}\"; load \"${GnuplotVarFile}\"; \
-				    load \"${my_dir}/random_tlength_at.gnuplot\"; quit" \
-				> ${ra_w_tlength_at_png}
-
-			ra_w_tspeed_tlength_png=${f%.*}-rw-ts_tl.png
-			echo "${f}: ${ra_w_tspeed_tlength_png}: Plot random write transfer speed - transfer length."
-			cat << EOF > ${GnuplotVarFile}
-set title "${Model} ${CapacityGB}G bytes,\\n\
-plot writes of random read/write\\n\
-${RandomRWMinBytesKi}Ki to ${RandomRWMaxBytesKi}Ki bytes per one write() call, \
-${DoDirectRandom}\\ntransfer speed - transfer length"
-pointcolor="#ff0000"
-set yrange [ ${RANDOM_TRANSFER_SPEED_MIN} : ${RANDOM_TRANSFER_SPEED_MAX} ] noreverse nowriteback
-EOF
-			gnuplot -e "log_file=\"${part_data_file}\"; load \"${GnuplotVarFile}\"; \
-				    load \"${my_dir}/random_tspeed_tlength.gnuplot\"; quit" \
-				> ${ra_w_tspeed_tlength_png}
-
-			rm "${part_data_file}"
-			rm "${ra_file}"
-		else
-			echo "${f}: Empty random access log."
-		fi
-	fi
-
-	if [[ "${DoReadFile}x" != "nx" ]]
+	if [[ "${DoReadFile}" != "n" ]]
 	then
 		cat << EOF > ${GnuplotVarFile}
 set title "${Model} ${CapacityGB}G bytes, sequential read\\n\
